User, Pothole,andFeedback endpoints working

+++ streetcare-backend/src/main/java/app_reporting_api/repository/FeedbackRepository.java
@@ -3,5 +3,5 @@ package app_reporting_api.repository;
 import app_reporting_api.model.FeedbackModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FeedbackRepository extends JpaRepository<FeedbackModel, Integer> {
+public interface FeedbackRepository extends JpaRepository<FeedbackModel, Long> {
 }
