+++ streetcare-backend/src/main/java/app_reporting_api/repository/PotholeRepository.java
@@ -3,5 +3,5 @@ package app_reporting_api.repository;
 import app_reporting_api.model.PotholeModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PotholeRepository extends JpaRepository<PotholeModel, Integer> {
+public interface PotholeRepository extends JpaRepository<PotholeModel, Long> {
 }
